QUAL-408 : Fixed package removal

+++ build/buildqual.sh
@@ -175,8 +175,11 @@ fi
 
 cp -r ${QUALDIR}/build/mps-builder/* ${MPSBUILDDIR}/
 
-sudo rm -f ${MPSBUILDDIR}/repo/packages/x86_64/nms-*.rpm
-sudo rm -f ${MPSBUILDDIR}/repo/packages/x86_64/tklabs_utils-*.rpm
+if [ "$NMS" == "YES" ]; then
+    sudo rm -f ${MPSBUILDDIR}/repo/packages/x86_64/nms-*.rpm
+    sudo rm -f ${MPSBUILDDIR}/repo/packages/x86_64/tklabs_utils-*.rpm
+fi
+
 sudo rm -f ${MPSBUILDDIR}/repo/packages/x86_64/qual-*.rpm
 sudo mv /tmp/tito/x86_64/* ${MPSBUILDDIR}/repo/packages/x86_64/
